Move constants into Zeroconf directory
Use hierarchical modules for Zeroconf
