+++ src/index.ts
@@ -1,12 +1,10 @@
+export * from './Network'
+export * from './Daemon'
 export * from './Client/Client'
 export * from './Client/Http2Client'
-export * from './Constants'
-export * from './Daemon'
-export * from './Network'
 export * from './Job/Job'
 export * from './Job/NormalJob'
 export * from './JobOrderer/HeapJobOrderer'
 export * from './JobOrderer/JobOrderer'
 export * from './JobOrderer/UnknownJobError'
-export * from './Zeroconf/ZeroconfClient'
-export * from './Zeroconf/ZeroconfDaemon'
+export * from './Zeroconf'
